Deprecate inclusion of main namespace module. Prefer inclusion of Verbalize::Action instead
