+++ lib/verbalize/error.rb
@@ -1,0 +1,3 @@
+module Verbalize
+  VerbalizeError = Class.new(StandardError)
+end
